qa_logic_backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.29
aiosqlite>=0.20.0
google-genai>=0.2.0
typing-extensions>=4.10.0
pytest>=8.0.0
openpyxl>=3.1.2
pytest-asyncio>=0.23.5
httpx>=0.27.0
pypdf>=4.2.0
python-docx>=1.1.2
python-multipart>=0.0.9
